requirements.txt
gradio
pytesseract
opencv-python-headless
numpy
Pillow
deep-translator